+++ src/middlewares/delay.ts
@@ -1,7 +1,12 @@
 import { Locals } from '../types/common.types';
 
-export default (_req, res, next) => {
-  const locals = res.locals as Locals;
-  const { delay } = locals.routeConfig || {};
-  !isNaN(delay as any) ? setTimeout(() => { next() }, delay) : next();
+export default (req, res, next) => {
+  const locals = res.locals as Locals || {};
+
+  const delay = locals.routeConfig?.delay || [].concat(req.query?._delay || 0)[0];
+  const _delay = parseInt(delay as any || 0, 10);
+
+  if (!_delay || isNaN(_delay)) return next();
+  
+  setTimeout(next, _delay);
 }