+++ src/utils/fetch.ts
@@ -155,11 +155,12 @@ export const getUrlData = async (request: AxiosRequestConfig): Promise<ValidType
     } else {
       response = await axios(request);
     }
+    const isImage = response.headers["content-type"]?.match(/image\/(jpeg|jpg|gif|png)$/gi)?.length > 0;
     fetchData = {
-      isImage: response.headers["content-type"]?.match(/image\/(jpeg|jpg|gif|png)$/gi)?.length > 0,
+      isImage,
       isError: false,
       statusCode: response.status,
-      headers: response.headers,
+      headers: { ...response.headers || {} },
       response: response.data
     };
   } catch (err: any) {